Refactor: OutputView로 수익률 출력 메서드 이동

+++ src/main/java/lotto/domain/LottoProfitRate.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import static lotto.utils.LottoConstant.LOTTO_PRICE;
 
 public class LottoProfitRate {
-    public double getProfitRate(double profit, int payment) {
-        return Math.round((profit / payment) * LOTTO_PRICE) / 10.0;
+    public double getProfitRate(long profit, int payment) {
+        return Math.round((profit / (double) payment) * LOTTO_PRICE) / 10.0;
     }
 }
